projet_location_propriete/requirements.txt
fastapi[standard]
mysql-connector-python
pydantic-settings
pydantic
pydantic[email]
pytest
pytest_bdd
requests
python-jose
starlette
python-multipart
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
bcrypt

# fastapi==0.115.2
# pydantic==2.9.2
# pydantic-settings==2.6.1
# mysql-connector-python==9.1.0